Refactoring existing code, local build success
