+++ btc-backend-commons/src/main/java/pl/btcgrouppl/btc/backend/commons/cqrs/models/annotations/CommandAnnotation.java
@@ -1,7 +1,5 @@
 package pl.btcgrouppl.btc.backend.commons.cqrs.models.annotations;
 
-import lombok.Data;
-
 import java.lang.annotation.ElementType;
 import java.lang.annotation.Retention;
 import java.lang.annotation.RetentionPolicy;
