+++ btc-backend-commons/src/main/java/pl/btcgrouppl/btc/backend/commons/integration/models/pojos/IntegrationMessage.java
@@ -1,6 +1,8 @@
 package pl.btcgrouppl.btc.backend.commons.integration.models.pojos;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NonNull;
 
 /**
  * Created by Sebastian Mekal <sebitg@gmail.com> on 01.06.15.
@@ -10,9 +12,13 @@ import lombok.Data;
  * </p>
  */
 @Data
+@AllArgsConstructor
 public class IntegrationMessage {
 
+    @NonNull
     private String messageClass;
+
+    @NonNull
     private Object body;
 
 }
