+++ btc-backend-commons/src/main/java/pl/btcgrouppl/btc/backend/commons/integration/models/pojos/IntegrationMessageEvent.java
@@ -1,6 +1,10 @@
 package pl.btcgrouppl.btc.backend.commons.integration.models.pojos;
 
+import lombok.AllArgsConstructor;
 import lombok.Data;
+import lombok.NonNull;
+import org.springframework.messaging.Message;
+import org.springframework.messaging.MessageHeaders;
 
 import java.util.Map;
 
@@ -12,8 +16,21 @@ import java.util.Map;
  * </p>
  */
 @Data
+@AllArgsConstructor
 public class IntegrationMessageEvent {
 
+    @NonNull
     private Map<String, Object> headers;
+    @NonNull
     private Object payload;
+
+    /**
+     * Creating from spring integration Message
+     * @param message
+     * @return IntegrationMessageEvent
+     */
+    public static IntegrationMessageEvent fromMessage(@NonNull Message<?> message) {
+        MessageHeaders headers = message.getHeaders();
+        return new IntegrationMessageEvent(headers, message.getPayload());
+    }
 }
