+++ btc-backend-commons/src/test/java/pl/btcgrouppl/btc/backend/commons/test/cqrs/CommandExecutorServiceImplTest.java
@@ -4,19 +4,17 @@ import org.junit.Test;
 import org.junit.runner.RunWith;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.beans.factory.annotation.Qualifier;
-import org.springframework.context.annotation.AnnotationConfigApplicationContext;
 import org.springframework.test.context.ContextConfiguration;
 import org.springframework.test.context.junit4.SpringJUnit4ClassRunner;
 import org.springframework.test.context.support.AnnotationConfigContextLoader;
 import pl.btcgrouppl.btc.backend.commons.BtcBackendCommonsSpringConfiguration;
 import pl.btcgrouppl.btc.backend.commons.cqrs.CommandExecutorService;
 import pl.btcgrouppl.btc.backend.commons.cqrs.CommandHandler;
-import pl.btcgrouppl.btc.backend.commons.cqrs.exceptions.CqrsException;
+import pl.btcgrouppl.btc.backend.commons.cqrs.models.exceptions.CqrsException;
 import pl.btcgrouppl.btc.backend.commons.test.util.cqrs.*;
-import rx.Observable;
 import rx.observables.BlockingObservable;
 
-import static org.junit.Assert.*;
+import static org.junit.Assert.assertEquals;
 import static tumbler.Tumbler.*;
 
 @RunWith(SpringJUnit4ClassRunner.class)
